+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Update to Python 3.14.4. Ensure that you update your uv first (`uv self update`).
 - Move evaluation of hydrodynamics to seperate file.
 - Also calculate discharge metrics at daily frequency if only hourly data is available.
+- Added auto-update for build methods.
 
 # v1.0.0b19
 - Add option for filling and raise out of bounds error for sample_from_map.